clean up db-clear
